refactor(dashboard): let SentinelApiClient to be an interface.
